fix(scraper): prevent integer overflow in revenue_estimate and fix deadlock retries
- Cast fee_amount to bigint before multiplication in compute_club_daily_stats
  to prevent PostgreSQL integer overflow on large VND fee × joined calculations
  (was crashing the KL and HCM market ingest with NumericValueOutOfRange)
- Add psycopg2.errors.DeadlockDetected to roster deadlock exception handler
  (DeadlockDetected was falling through to generic except, bypassing retry)
- Cap deadlock/connection retries at 2 with exponential back-off via _retry param
  to prevent unbounded recursion on repeated deadlocks

+++ ingest.py
@@ -625,8 +625,8 @@ def compute_club_daily_stats(cur):
             CASE WHEN SUM(s.max_players) > 0
                  THEN ROUND(CAST(COALESCE(SUM(ds.joined), 0) AS numeric) / SUM(s.max_players), 3)
                  ELSE 0 END AS avg_fill_rate,
-            ROUND(AVG(s.fee_amount)) AS avg_fee,
-            COALESCE(SUM(ds.joined * s.fee_amount), 0) AS revenue_estimate
+            ROUND(AVG(s.fee_amount::bigint)) AS avg_fee,
+            COALESCE(SUM(ds.joined * s.fee_amount::bigint), 0) AS revenue_estimate
         FROM sessions s
         LEFT JOIN LATERAL (
             SELECT joined, waitlisted FROM daily_snapshots
